# Optimize distributed rectangular domain creation