@ronawho 2022-11-14
I was a little surprised that this didn't require any comm count .good file updates, but it looks like all the tests create the domain outside any tracked region (i.e. the distribution tests create domains in the untracked `driver_domains` module)

We will see the change in performance and comm count graphs, but could add a correctness test if that seems valuable.